backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
google-generativeai==0.4.1
scikit-learn==1.4.1.post1
clarifai-grpc==10.1.1
firebase-admin==6.4.0
python-dotenv==1.0.1
